requirements.txt
streamlit
pycryptodome
pypdf2
Pillow
python-docx
pandas
